+++ spec/Component/ContentElement/AbstractContentElementSpec.php
@@ -40,10 +40,10 @@ class AbstractContentElementSpec extends ObjectBehavior
     function let(EngineInterface $templateEngine, RequestScopeMatcher $requestScopeMatcher)
     {
         $this->modelData = [
-            'type' => 'test',
-            'headline' => serialize(['unit' => 'h1', 'value' => 'test']),
-            'id'   => 1,
-            'cssID' => serialize(['', '']),
+            'type'      => 'test',
+            'headline'  => serialize(['unit' => 'h1', 'value' => 'test']),
+            'id'        => 1,
+            'cssID'     => serialize(['', '']),
             'customTpl' => 'custom_tpl'
         ];