+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 
 ### Added
 
+ - Introduce interface `Netzmacht\Contao\Toolkit\Dca\DcaManager`
  - Add alias `netzmacht.contao_toolkit.csrf.token_manager` which refers to the token manager used in Contao
  - Add CsrfTokenProvider to simplify access to the request token
  - Add `Netzmacht\Contao\Toolkit\View\Assets\HtmlPageAssetsManager` interface which extends the AssetsManager